updated gitignore.
FUNFACT: This commit was supposed to be the commit where I first added
v1 of this scraper. But I misunderstood what I was doing
with git and accidentally removed the entire local repo.

It was devastating.

So, lesson learned. Commit Early Commit Often. I am now going to rewrite
the project.

 Changes to be committed:
	new file:   .gitignore